refactor(store): merge suite.storage and suite.search into suite.store
Rename the suite.storage package to suite.store and fold the search
module into it: base_index.py becomes search_store.py and its BaseIndex
class becomes SearchStore, aligning with the base_store/data_store/
blob_store naming. get_search_base_path now lives in suite.store, and
the storage/search loggers are unified under "suite.store".

Repoint all importers (mail storage/search wrappers, jmap_account,
store_entry, and the cache-migration patch) at the merged package.
